+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleaed
+
+### Added
+- A121: Calibration reuse example
+
 ## v5.4.1
 
 ### Added